# csc_openapi_2.1.0.1.yaml
openapi: 3.0.0
info:
  version: 2.1.0.1
  title: CSC Api
  description:
    This is a CSC v2 openapi specification based on the _CSC V2.1 API Specification_ (CSC Spec.).
  contact:
    name: Cloud Signature Consortium
    url: https://cloudsignatureconsortium.org/
paths:
  /v2/info:
    post:
      tags:
        - info
      summary: Service information
      description: >-
        Returns information about the remote service and the list of the API methods it supports. 
        This method SHALL be implemented by any remote service conforming to this specification.
      operationId: info
      requestBody:
        $ref: '#/components/requestBodies/input-info'
      responses:
        '200':
          $ref: '#/components/responses/output-info'
        '400':
          $ref: '#/components/responses/invalidRequestResponse'
        '401':
          $ref: '#/components/responses/unauthorizedResponse'
        '429':
          $ref: '#/components/responses/tooManyRequestsResponse'
        '500':
          $ref: '#/components/responses/internalServerErrorResponse'
        '501':
          $ref: '#/components/responses/notImplementedResponse'
        '503':
          $ref: '#/components/responses/serviceUnavailableResponse'
  /v2/auth/login:
    post:
      tags:
        - auth
      summary: Get access token
      security:
        - BasicAuth: []
        - BearerAuth: []
      description: >-
        Obtain an access token for service authorization from the Remote Service
        using HTTP Basic Authentication or HTTP Digest authentication, as defined 
        in RFC 7235 [2], using the userID and password assigned to the user.
      operationId: authLogin
      requestBody:
        $ref: '#/components/requestBodies/input-auth-login'
      responses:
        '200':
          $ref: '#/components/responses/output-auth-login'
        '400':
          $ref: '#/components/responses/invalidRequestResponse'
        '401':
          $ref: '#/components/responses/unauthorizedResponse'
        '429':
          $ref: '#/components/responses/tooManyRequestsResponse'
        '500':
          $ref: '#/components/responses/internalServerErrorResponse'
        '501':
          $ref: '#/components/responses/notImplementedResponse'
        '503':
          $ref: '#/components/responses/serviceUnavailableResponse'
  /v2/auth/revoke:
    post:
      tags:
        - auth
      summary: Revoke oAuth2 access token
      security:
        - BearerAuth: []
      description: >-
        Revoke a service access token or refresh token that was obtained from
        the Remote Service or an associated Authorization server.
      operationId: authRevoke
      requestBody:
        $ref: '#/components/requestBodies/input-auth-revoke'
      responses:
        '204':
          description: >-
            Operation successfully executed
        '400':
          $ref: '#/components/responses/invalidRequestResponse'
        '401':
          $ref: '#/components/responses/unauthorizedResponse'
        '429':
          $ref: '#/components/responses/tooManyRequestsResponse'
        '500':
          $ref: '#/components/responses/internalServerErrorResponse'
        '501':
          $ref: '#/components/responses/notImplementedResponse'
        '503':
          $ref: '#/components/responses/serviceUnavailableResponse'
  /v2/credentials/list:
    post:
      tags:
        - credentials
      summary: User credentials
      security:
        - BearerAuth: []
      description: Returns the list of credentials associated with a user identifier
      operationId: credentialsList
      requestBody:
        $ref: '#/components/requestBodies/input-credentials-list'
      responses:
        '200':
          $ref: '#/components/responses/output-credentials-list'
        '400':
          $ref: '#/components/responses/invalidRequestResponse'
        '401':
          $ref: '#/components/responses/unauthorizedResponse'
        '429':
          $ref: '#/components/responses/tooManyRequestsResponse'
        '500':
          $ref: '#/components/responses/internalServerErrorResponse'
        '501':
          $ref: '#/components/responses/notImplementedResponse'
        '503':
          $ref: '#/components/responses/serviceUnavailableResponse'
  /v2/credentials/info:
    post:
      tags:
        - credentials
      summary: User credential information
      security:
        - BearerAuth: []
      description: >-
        Retrieve the credential and return the main identity information and the
        public key certificate or the certificate chain associated to it.
      operationId: credentialsInfo
      requestBody:
        $ref: '#/components/requestBodies/input-credentials-info'
      responses:
        '200':
          $ref: '#/components/responses/output-credentials-info'
        '400':
          $ref: '#/components/responses/invalidRequestResponse'
        '401':
          $ref: '#/components/responses/unauthorizedResponse'
        '429':
          $ref: '#/components/responses/tooManyRequestsResponse'
        '500':
          $ref: '#/components/responses/internalServerErrorResponse'
        '501':
          $ref: '#/components/responses/notImplementedResponse'
        '503':
          $ref: '#/components/responses/serviceUnavailableResponse'
  /v2/credentials/authorize:
    post:
      tags:
        - credentials
      summary: Credentials access authorization
      security:
        - BearerAuth: []
      description: >-
        Authorize the access to the credential for remote signing, according to the 
        authorization mechanisms associated to it. This method returns the Signature 
        Activation Data (SAD) required to authorize the signatures/signHash method, 
        as defined in signatures/signHash or signatures/signDoc method, as defined 
        in signatures/signDoc.
      operationId: credentialsAuthorize
      requestBody:
        $ref: '#/components/requestBodies/input-credentials-authorize'
      responses:
        '200':
          $ref: '#/components/responses/output-credentials-authorize-SAD'
        '202':
          $ref: '#/components/responses/output-credentials-authorize-handle'
        '400':
          $ref: '#/components/responses/invalidAuthorizeResponse'
        '401':
          $ref: '#/components/responses/unauthorizedResponse'
        '429':
          $ref: '#/components/responses/tooManyRequestsResponse'
        '500':
          $ref: '#/components/responses/internalServerErrorResponse'
        '501':
          $ref: '#/components/responses/notImplementedResponse'
        '503':
          $ref: '#/components/responses/serviceUnavailableResponse'
...
  /v2/signatures/signHash:
    post:
      tags:
        - signatures
      summary: >-
          Hash Signing
      security:
        - BearerAuth: []
      description:  Calculate the remote digital signature of one or multiple hash values provided as an input.
      operationId: signaturesSignHash
      requestBody:
        $ref: '#/components/requestBodies/input-signatures-signhash'
      responses:
        '200':
          $ref: '#/components/responses/output-signatures-signhash'
        '400':
          $ref: '#/components/responses/invalidSignaturesSignHashRequestResponse'
        '401':
          $ref: '#/components/responses/unauthorizedResponse'
        '429':
          $ref: '#/components/responses/tooManyRequestsResponse'
        '500':
          $ref: '#/components/responses/internalServerErrorResponse'
        '501':
          $ref: '#/components/responses/notImplementedResponse'
        '503':
          $ref: '#/components/responses/serviceUnavailableResponse'          
...
servers:
  - url: https://{host}:{port}/xxx/csc/
components:
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
    BearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    lang:
      type: string
      description: Preferred language of the response to the remote service, specified according to RFC 5646
...
  requestBodies:
    input-info:
      content:
        application/json:
          schema:
            description: Input parameters for info method
            properties:
              lang:
                $ref: '#/components/schemas/lang'
            example:
              lang: en-US
    input-auth-login:
      content:
        application/json:
          schema:
            description: Input parameters for auth/login method
            properties:
              refresh_token:
                $ref: '#/components/schemas/refresh_token'
              rememberMe:
                $ref: '#/components/schemas/rememberMe'
              clientData:
                $ref: '#/components/schemas/clientData'
            example:
              rememberMe: true
    input-auth-revoke:
      content:
        application/json:
          schema:
            description: Input parameters for auth/revoke method
            required:
              - token
            properties:
              token:
                $ref: '#/components/schemas/token'
              token_type_hint:
                $ref: '#/components/schemas/token_type_hint'
              clientData:
                $ref: '#/components/schemas/clientData'
            example:
              token: _TiHRG-bA-H3XlFQZ3ndFhkXf9P24/CKN69L8gdSYp5_pw
              token_type_hint: refresh_token
              clientData: '12345678'
    input-credentials-list:
      content:
        application/json:
          schema:
            description: Input parameters for credentials/list method
            properties:
              userID:
                $ref: '#/components/schemas/userID'
              credentialInfo:
                $ref: "#/components/schemas/credentialInfo"
              certificates:
                $ref: "#/components/schemas/certificates"
              certInfo:
                $ref: "#/components/schemas/certInfo"
              authInfo:
                $ref: "#/components/schemas/authInfo"
              onlyValid:
                $ref: "#/components/schemas/onlyValid"
              lang:
                $ref: '#/components/schemas/lang'
              clientData:
                $ref: '#/components/schemas/clientData'
            example:
              credentialInfo: true
              certificates: "chain"
              certInfo: true
              authInfo: true
    input-credentials-info:
      content:
        application/json:
          schema:
            description: Input parameters for credentials/info method
            required:
              - credentialID
            properties:
              credentialID:
                $ref: '#/components/schemas/credentialID'
              certificates:
                $ref: '#/components/schemas/certificates'
              certInfo:
                $ref: '#/components/schemas/certInfo'
              authInfo:
                $ref: '#/components/schemas/authInfo'
              lang:
                $ref: '#/components/schemas/lang'
              clientData:
                $ref: '#/components/schemas/clientData'
            example:
              credentialID: GX0112348
              certificates: chain
              certInfo: true
              authInfo: true
    input-credentials-authorize:
      content:
        application/json:
          schema:
            description: Input parameters for credentials/authorize method
            required:
              - credentialID
              - numSignatures
            properties:
              credentialID:
                $ref: '#/components/schemas/credentialID'
              numSignatures:
                $ref: '#/components/schemas/numSignatures'
              hashes:
                $ref: '#/components/schemas/hashes'
              hashAlgorithmOID:
                $ref: "#/components/schemas/hashAlgorithmOID"
              authData:
                type: array
                items:
                  $ref: "#/components/schemas/authData"
              description:
                $ref: '#/components/schemas/description'
              clientData:
                $ref: '#/components/schemas/clientData'
            example:
              credentialID: GX0112348
              numSignatures: 2
              hashes: 
              - sTOgwOm+474gFj0q0x1iSNspKqbcse4IeiqlDg/HWuI=
              - c1RPZ3dPbSs0NzRnRmowcTB4MWlTTnNwS3FiY3NlNEllaXFsRGcvSFd1ST0=
              hashAlgorithmOID: 2.16.840.1.101.3.4.2.1,
              authData:
              - id: PIN
                value: '123456'
              - id: OTP
                value: '738496'
              clientData: '12345678'
...
    input-signatures-signhash:
      content:
        application/json:
          schema:
            description: Input parameters for signatures/signHash method
            required:
              - credentialID
              - hash
              - signAlgo
            properties:
              credentialID:
                $ref: '#/components/schemas/credentialID'
              SAD:
                $ref: '#/components/schemas/SAD'
              hashes:
                $ref: '#/components/schemas/hashes'
              hashAlgorithmOID:
                $ref: '#/components/schemas/hashAlgorithmOID'
              signAlgo:
                $ref: '#/components/schemas/signAlgo'
              signAlgoParams:
                $ref: '#/components/schemas/signAlgoParams'
              operationMode:
                $ref: '#/components/schemas/operationMode'
              validity_period:
                $ref: '#/components/schemas/validity_period'
              response_uri:
                $ref: '#/components/schemas/response_uri'
              clientData:
                $ref: '#/components/schemas/clientData'
            example:
              credentialID: GX0112348
              SAD: _TiHRG-bAH3XlFQZ3ndFhkXf9P24/CKN69L8gdSYp5_pw
              hashes:
                - sTOgwOm+474gFj0q0x1iSNspKqbcse4IeiqlDg/HWuI=
                - c1RPZ3dPbSs0NzRnRmowcTB4MWlTTnNwS3FiY3NlNEllaXFsRGcvSFd1ST0=
              hashAlgorithmOID: 2.16.840.1.101.3.4.2.1
              signAlgo: 1.2.840.113549.1.1.1
              clientData: '12345678'
...
  responses:
    output-info:
      description: Info response object
      content:
        application/json:
          schema:
            required:
              - specs
              - name
              - logo
              - region
              - lang
              - description
              - authType
              - methods
              - signAlgorithms
              - signature_formats
              - conformance_levels
            properties:
              specs:
                $ref: '#/components/schemas/specs'
              name:
                $ref: '#/components/schemas/name'
              logo:
                $ref: '#/components/schemas/logo'
              region:
                $ref: '#/components/schemas/region'
              lang:
                $ref: '#/components/schemas/lang'
              description:
                $ref: '#/components/schemas/description'
              authType:
                $ref: '#/components/schemas/authType'
              oauth2Servers:
                $ref: '#/components/schemas/oauth2Servers'
              oauth2:
                $ref: '#/components/schemas/oauth2'
              oauth2Issuer:
                $ref: '#/components/schemas/oauth2Issuer'
              supportsRar:
                type: boolean
                description: >-
                  If this parameter is present and either “oauth2” or “oauth2issuer” is present, this parameters SHALL be “true” if the authorization server supports “authorization_details” as documented in oauth2/authorize and “false” if it does not. If this parameter is not present and either “oauth2” or “oauth2issuer” is present, the default value is "false"
              supportedHashTypes:
                type: array
                items:
                  type: string
              asynchronousOperationMode:
                $ref: '#/components/schemas/asynchronousOperationMode'
              methods:
                $ref: '#/components/schemas/methods'
              validationInfo:
                type: boolean
                description: >-  
                  This parameter SHALL be “true” if the remote signing server supports the “validationInfo” response parameter of the method signatures/signDoc in not mandatory cases. An omitted parameter or the value “false” indicates that the remote signing server does not support “validationInfo” in those cases.
                default: false
              signAlgorithms:
                $ref: '#/components/schemas/signAlgorithms'
              documentTypes:
                $ref: '#/components/schemas/documentTypes'
              signature_formats:
                $ref: '#/components/schemas/signature_formats'
              conformance_levels:
                $ref: '#/components/schemas/conformance_levels'
            example:
              specs: 2.0.0.0
              name: ACME Trust Services
              logo: https://service.domain.org/images/logo.png
              region: IT
              lang: en-US
              description: An efficient remote signature service
              authType:
                - basic
                - oauth2code
              oauth2: https://www.domain.org/
              methods:
                - auth/login
                - auth/revoke
                - credentials/list
                - credentials/info
                - credentials/authorize
                - signatures/signOTP
                - signatures/signHash
                - signatures/signDoc


              signAlgorithms:
                - algorithm: 1.2.840.10045.4.3.2
                - algorithm: 1.2.840.113549.1.1.1
                - algorithm: 1.2.840.113549.1.1.10
              documentTypes:
                - sod
                - sfd
              signature_formats:
                formats: 
                  - C
                  - X
                  - P
                envelope_properties:
                  - - Detached
                    - Attached
                    - Parallel
                  - - Enveloped
                    - Enveloping
                    - Detached
                  - - Certification
                    - Revision
                allowMix: true
              conformance_levels:
                - Ades-B-B
                - Ades-B-T
    output-auth-login:
      description: auth/login response object
      content:
        application/json:
          schema:
            required:
              - access_token
            properties:
              access_token:
                $ref: '#/components/schemas/access_token'
              refresh_token:
                $ref: '#/components/schemas/refresh_token'
              expires_in:
                $ref: '#/components/schemas/expires_in'
            example:
              access_token: 4/CKN69L8gdSYp5_pwH3XlFQZ3ndFhkXf9P2_TiHRG-bA
              refresh_token: _TiHRG-bA-H3XlFQZ3ndFhkXf9P24/CKN69L8gdSYp5_pw
              expires_in: 3600
    output-credentials-list:
      description: credentials/list response object
      content:
        application/json:
          schema:
            required:
              - credentialIDs
            properties:
              credentialIDs:
                $ref: '#/components/schemas/credentialIDs'
              credentialInfos:
                $ref: '#/components/schemas/credentialInfos'
              onlyValid:
                $ref: '#/components/schemas/onlyValid'
            example:
              credentialIDs:
                - GX0112348
                - HX0224685
              credentialInfos:
              - credentialID: GX0112348
                key:
                  status: enabled
                  algo:
                  - 1.2.840.113549.1.1.11
                  - 1.2.840.113549.1.1.10
                  len: 2048
                cert:
                  status: valid
                  certificates:
                  - "<Base64-encoded_X.509_end_entity_certificate>"
                  - "<Base64-encoded_X.509_intermediate_CA_certificate>"
                  - "<Base64-encoded_X.509_root_CA_certificate>"
                  issuerDN: "<X.500_issuer_DN_printable_string>"
                  serialNumber: 5AAC41CD8FA22B953640
                  subjectDN: "<X.500_subject_DN_printable_string>"
                  validFrom: 20200101100000Z
                  validTo: 20230101095959Z
                auth:
                  mode: explicit
                  expression: PIN AND OTP
                  objects:
                  - type: Password
                    id: PIN
                    format: N
                    label: PIN
                    description: Please enter the signature PIN
                  - type: Password
                    id: OTP
                    format: N
                    generator: totp
                    label: Mobile OTP
                    description: Please enter the 6 digit code you received by SMS
                multisign: 5
                lang: en-US
    output-credentials-info:
      description: credentials/info response object
      content:
        application/json:
          schema:
            required:
              - credentialID
              - key
              - authMode
              - PIN
              - OTP
              - multisign
              - lang
            properties:
              description:
                $ref: '#/components/schemas/description'
              signatureQualifier:
                $ref: '#/components/schemas/signatureQualifier'
              key:
                $ref: '#/components/schemas/key'
              cert:
                $ref: '#/components/schemas/cert'
              auth:
                $ref: '#/components/schemas/auth'
              SCAL:
                $ref: '#/components/schemas/SCAL'
              multisign:
                $ref: '#/components/schemas/multisign'
              lang:
                $ref: '#/components/schemas/lang'
            example:
              key:
                status: enabled
                algo:
                  - 1.2.840.113549.1.1.1
                  - 0.4.0.127.0.7.1.1.4.1.3
                len: 2048
              cert:
                status: valid
                certificates:
                  - <Base64-encoded_X.509_end_entity_certificate>
                  - <Base64-encoded_X.509_intermediate_CA_certificate>
                  - <Base64-encoded_X.509_root_CA_certificate>
                issuerDN: <X.500_issuer_DN_printable_string>
                serialNumber: 5AAC41CD8FA22B953640
                subjectDN: <X.500_subject_DN_printable_string>
                validFrom: 20180101100000Z
                validTo: 20190101095959Z
                authMode: explicit
                multisign: 5
                lang: en-US
    output-credentials-authorize-SAD:
      description: credentials/authorize response object
      content:
        application/json:
          schema:
            required:
              - SAD
            properties:
              SAD:
                $ref: '#/components/schemas/SAD'
              expiresIn:
                $ref: '#/components/schemas/expiresIn'
            example:
              SAD: 1/UsHDJ98349h9fgh9348hKKHDkHWVkl/8hsAW5usc8_5=
    output-credentials-authorize-handle:
      description: credentials/authorizeHandle response object
      content:
        application/json:
          schema:
            required:
              - handle
            properties:
              handle:
                $ref: '#/components/schemas/handle'
            example:
              handle: "878287f37b2bv293bv2bv237bv297bvbv"
...
    output-signatures-signhash:
      description: signatures/signHash response object
      content:
        application/json:
          schema:
            properties:
                signatures:
                  $ref: '#/components/schemas/signatures'
                responseID:
                  $ref: '#/components/schemas/responseID'
            example:
              signatures:
                - KedJuTob5gtvYx9qM3k3gm7kbLBwVbEQRl26S2tmXjqNND7MRGtoew==
                - Idhef7xzgtvYx9qM3k3gm7kbLBwVbE98239S2tm8hUh85KKsfdowel==
...
    invalidRequestResponse:
      description: The request has failed due to invalid or missing required parameters. Error string is "invalid_request".
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/errorResponse'
    invalidAuthorizeResponse:
      description: The request has failed due to invalid authentication data, invalid or missing required parameters. Error string is either "invalid_request" or "invalid_authentication_data".
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/errorResponse'
...
    unauthorizedResponse:
      description: |
        The request has failed due to bad or expired authorization token being used. Error string is one of the following
        - "unauthorized_client"
        - "access_denied"
        - "expired_token"
        - "invalid_token"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/errorResponse'
    tooManyRequestsResponse:
      description: The request has failed due to rate limiting. Error string is "temporarily_unavailable".
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/errorResponse'
    internalServerErrorResponse:
      description: The request has failed due to an unexpected condition. Error string is "server_error".
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/errorResponse'
    notImplementedResponse:
      description: The request has failed due to an unimplemented method being requested. Error string is "not_implemented".
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/errorResponse'
    serviceUnavailableResponse:
      description: The request has failed due to temporary overloading or maintenance conditions. Error string is "temporarily_unavailable".
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/errorResponse'